step3 refactor : cardrule로 패키지명 변경
