+++ src/main/kotlin/blackjack/domain/cardrule/PlayingCardsRule.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.rule
+package blackjack.domain.cardrule
 
 import blackjack.domain.PlayingCards
